--- src/classes/filterpriority.js
@@ -1,8 +1,0 @@
-import Store from './store';
-
-let todos = Store.getTodos();
-const filterByPriority = (priority) => {
-    let todos = todos.map(todo => todo.priority = priority);
-}
-
-export {filterByPriority as default};
