--- src/classes/project.js
@@ -1,2 +1,0 @@
-
-
