--- src/helpers/rootpath.js
@@ -1,5 +1,0 @@
-const rootPath = () => {
-    return '/home/henok/Documents/microverse/Javascript/todo-list';        
-}
-
-export { rootPath as default };
